ArduinoConnector/requirements.txt
fastapi
uvicorn
pydantic
pyserial